Organize and update site
A lot of things happened since the last complete modification of the
site:
- GitHub Pages switched to Jekyll 3.
- The new Jekyll changed some things and added others.
- I learned to use Bundler.
- Some plugins are now allowed for use with GitHub.
- And much more, which I'll skip for now.

Because of that reasons, there was a need to refactor the entire site,
because the initial one had redundant content, bad directory and assets
structure, poor usage of data files, and even unorganized configuration,
adding to it the results of the merge of `gh-pages` branch with `master`
because the other wasn't needed anymore and Jekyll was mature enough to
handle things better.

With this commit, there are too many changes to describe in here, but in
summary:
- Data from `index.html` was moved to `_data/tools`.
- Asset urls now depend on environment and are in `_data/cdn`.
- Simplify `README.md` and concentrate in the webpage.
- Update `LICENSE` years everywhere.
- Simplify `_includes/vars.html` by using plugins.
- Removed unnecessary files.

Hope this commit serves as a base for future Jekyll sites by me.